feat(transfersync): auto-update nameservers to the requested ones after successful transfer
BREAKING CHANGE: WHMCS provides specifying new Nameservers within domain transfer process. This is
not compliant with EPP standard of the most registries and not working for the most TLDs using the
HEXONET API. Therefore we care about updating on WHMCS-side. As this is a completely different
module behavior, we trigger a new major release.