feat(dns management): create DNS Zone after paid invoice (post-registration addon activation)
DNS RR Updates might fail when creating the DNSZone (when not existing) and updating it right after.
The creation of the DNSZone might still be pending/being in progress and the DNS RR Update will
therefore fail. This is a delicate issue just because WHMCS is only showing a common error message
not allowing the end-customer (nor the reseller) identifying the root cause. That's why we reviewed
the entire process of the DNS Management Addon Activation on our end while WHMCS isn't providing a
useful registrar module function to integrate.